mysql-python/pyproject.toml
[tool.poetry]
name = "mysql-python"
version = "0.1.0"
description = ""
authors = ["xerio1001 <dariovanhasselt@hotmail.be>"]

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^8.0.28"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
